JPA first try.

+++ src/main/java/org/example/JdbcDemo.java
@@ -32,7 +32,7 @@ public class JdbcDemo {
             ResultSet rs = statement.executeQuery("SELECT * FROM Person");
             ResultSetMetaData metaData = rs.getMetaData();
             while (rs.next()) {
-                int id = rs.getInt("id");
+                long id = rs.getInt("id");
                 String name = rs.getString("name");
                 int age = rs.getInt("age");
